Move download Gerber functionality to separate file +
- Change inches to mm
- Add comments on top of the gerber file
- Add gerber creation timestamp
- Add comments with Gerber best practices
